Removed the unnecessary logger from the Calculator.cs class, and changed it's methods' parameters to take 'double' numbers instead of 'int'. ICalculator.cs interface changed as necessary.

+++ SimpleCalculator/Service/ICalculator.cs
@@ -2,8 +2,8 @@ namespace SimpleCalculator.Service;
 
 public interface ICalculator
 {
-    double Add(int a, int b);
-    double Subtract(int a, int b);
-    double Multiply(int a, int b);
-    double Divide(int a, int b);
+    double Add(double a, double b);
+    double Subtract(double a, double b);
+    double Multiply(double a, double b);
+    double Divide(double a, double b);
 }
